interface/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
aiohttp>=3.10.0
sse-starlette>=2.1.0
httpx>=0.27.0
pydantic>=2.0.0
python-multipart>=0.0.20
sentence-transformers>=5.2.0
faiss-cpu>=1.13.0
redis>=7.0.0
